+++ internal/api/server_test.go
@@ -1,23 +1,15 @@
 package api
 
 import (
-	"bytes"
 	"context"
-	"crypto/rand"
-	"crypto/sha256"
-	"encoding/hex"
 	"encoding/json"
-	"mime/multipart"
 	"net/http"
 	"net/http/httptest"
 	"path/filepath"
-	"strings"
 	"testing"
 	"time"
 
-	"github.com/wiebe-xyz/bugbarn/internal/auth"
 	"github.com/wiebe-xyz/bugbarn/internal/event"
-	"github.com/wiebe-xyz/bugbarn/internal/ingest"
 	"github.com/wiebe-xyz/bugbarn/internal/storage"
 	"github.com/wiebe-xyz/bugbarn/internal/worker"
 )
@@ -207,364 +199,6 @@ func TestServeHTTPQueryEndpoints(t *testing.T) {
 		}
 		if got, want := response.Events[2].ID, eventA1.ID; got != want {
 			t.Fatalf("unexpected third live event: got %q want %q", got, want)
-		}
-	})
-}
-
-func TestServeHTTPStatefulEndpoints(t *testing.T) {
-	t.Parallel()
-
-	store := mustOpenStore(t)
-	defer store.Close()
-
-	base := time.Date(2026, 4, 15, 12, 0, 0, 0, time.UTC)
-	issue, _ := mustPersistProcessedEvent(t, store, worker.ProcessedEvent{
-		Fingerprint: "fingerprint-stateful",
-		Event: event.Event{
-			ReceivedAt: base,
-			ObservedAt: base,
-			Severity:   "ERROR",
-			Message:    "stateful error",
-			Exception: event.Exception{
-				Type:    "panic",
-				Message: "stateful error",
-			},
-		},
-	})
-
-	server := NewServer(nil, store, nil)
-
-	t.Run("resolve and reopen issue", func(t *testing.T) {
-		rr := httptest.NewRecorder()
-		req := httptest.NewRequest(http.MethodPost, "/api/v1/issues/"+issue.ID+"/resolve", nil)
-		server.ServeHTTP(rr, req)
-		if rr.Code != http.StatusOK {
-			t.Fatalf("unexpected resolve status: got %d want %d body=%q", rr.Code, http.StatusOK, rr.Body.String())
-		}
-
-		rr = httptest.NewRecorder()
-		req = httptest.NewRequest(http.MethodPost, "/api/v1/issues/"+issue.ID+"/reopen", nil)
-		server.ServeHTTP(rr, req)
-		if rr.Code != http.StatusOK {
-			t.Fatalf("unexpected reopen status: got %d want %d body=%q", rr.Code, http.StatusOK, rr.Body.String())
-		}
-	})
-
-	t.Run("release list and create", func(t *testing.T) {
-		body := bytes.NewBufferString(`{"name":"v1.2.3","environment":"staging","observedAt":"2026-04-15T12:00:00Z","version":"1.2.3","commitSha":"abc123","url":"https://example.com","notes":"deploy"}`)
-		rr := httptest.NewRecorder()
-		req := httptest.NewRequest(http.MethodPost, "/api/v1/releases", body)
-		req.Header.Set("Content-Type", "application/json")
-		server.ServeHTTP(rr, req)
-		if rr.Code != http.StatusOK {
-			t.Fatalf("unexpected release create status: got %d want %d body=%q", rr.Code, http.StatusOK, rr.Body.String())
-		}
-
-		rr = httptest.NewRecorder()
-		req = httptest.NewRequest(http.MethodGet, "/api/v1/releases", nil)
-		server.ServeHTTP(rr, req)
-		if rr.Code != http.StatusOK {
-			t.Fatalf("unexpected release list status: got %d want %d body=%q", rr.Code, http.StatusOK, rr.Body.String())
-		}
-		var response struct {
-			Releases []storage.Release `json:"releases"`
-		}
-		decodeResponse(t, rr, &response)
-		if len(response.Releases) == 0 {
-			t.Fatal("expected release in list")
-		}
-	})
-
-	t.Run("alert create and settings", func(t *testing.T) {
-		body := bytes.NewBufferString(`{"name":"High errors","enabled":true,"severity":"error","condition":"count>10","query":"issue:123","target":"slack"}`)
-		rr := httptest.NewRecorder()
-		req := httptest.NewRequest(http.MethodPost, "/api/v1/alerts", body)
-		req.Header.Set("Content-Type", "application/json")
-		server.ServeHTTP(rr, req)
-		if rr.Code != http.StatusOK {
-			t.Fatalf("unexpected alert create status: got %d want %d body=%q", rr.Code, http.StatusOK, rr.Body.String())
-		}
-
-		rr = httptest.NewRecorder()
-		req = httptest.NewRequest(http.MethodPut, "/api/v1/settings", bytes.NewBufferString(`{"displayName":"BugBarn","timezone":"UTC","defaultEnvironment":"staging","liveWindowMinutes":15,"stacktraceContextLines":3}`))
-		req.Header.Set("Content-Type", "application/json")
-		server.ServeHTTP(rr, req)
-		if rr.Code != http.StatusOK {
-			t.Fatalf("unexpected settings status: got %d want %d body=%q", rr.Code, http.StatusOK, rr.Body.String())
-		}
-	})
-
-	t.Run("source map upload", func(t *testing.T) {
-		var body bytes.Buffer
-		writer := multipart.NewWriter(&body)
-		_ = writer.WriteField("release", "1.2.3")
-		_ = writer.WriteField("bundle_url", "https://example.com/app.js")
-		_ = writer.WriteField("source_map_name", "app.js.map")
-		part, err := writer.CreateFormFile("source_map", "app.js.map")
-		if err != nil {
-			t.Fatal(err)
-		}
-		if _, err := part.Write([]byte(`{"version":3}`)); err != nil {
-			t.Fatal(err)
-		}
-		if err := writer.Close(); err != nil {
-			t.Fatal(err)
-		}
-
-		rr := httptest.NewRecorder()
-		req := httptest.NewRequest(http.MethodPost, "/api/v1/source-maps", &body)
-		req.Header.Set("Content-Type", writer.FormDataContentType())
-		server.ServeHTTP(rr, req)
-		if rr.Code != http.StatusAccepted {
-			t.Fatalf("unexpected source map status: got %d want %d body=%q", rr.Code, http.StatusAccepted, rr.Body.String())
-		}
-	})
-}
-
-func TestServeHTTPUserAuthentication(t *testing.T) {
-	t.Parallel()
-
-	store := mustOpenStore(t)
-	defer store.Close()
-
-	userAuth, err := auth.NewUserAuthenticator("admin", "change-me", "")
-	if err != nil {
-		t.Fatal(err)
-	}
-	sessions := auth.NewSessionManager("test-secret", time.Hour)
-	server := NewServerWithAuth(nil, store, userAuth, sessions, nil, nil)
-
-	t.Run("query endpoints require session", func(t *testing.T) {
-		rr := httptest.NewRecorder()
-		req := httptest.NewRequest(http.MethodGet, "/api/v1/issues", nil)
-
-		server.ServeHTTP(rr, req)
-
-		if rr.Code != http.StatusUnauthorized {
-			t.Fatalf("unexpected status: got %d want %d", rr.Code, http.StatusUnauthorized)
-		}
-	})
-
-	var session *http.Cookie
-	t.Run("login sets session cookie", func(t *testing.T) {
-		rr := httptest.NewRecorder()
-		req := httptest.NewRequest(http.MethodPost, "/api/v1/login", strings.NewReader(`{"username":"admin","password":"change-me"}`))
-		req.Header.Set("Content-Type", "application/json")
-
-		server.ServeHTTP(rr, req)
-
-		if rr.Code != http.StatusOK {
-			t.Fatalf("unexpected status: got %d want %d body=%q", rr.Code, http.StatusOK, rr.Body.String())
-		}
-		for _, cookie := range rr.Result().Cookies() {
-			if cookie.Name == "bugbarn_session" {
-				session = cookie
-			}
-		}
-		if session == nil || session.Value == "" {
-			t.Fatal("expected session cookie")
-		}
-	})
-
-	t.Run("session can access query endpoint", func(t *testing.T) {
-		rr := httptest.NewRecorder()
-		req := httptest.NewRequest(http.MethodGet, "/api/v1/issues", nil)
-		req.AddCookie(session)
-
-		server.ServeHTTP(rr, req)
-
-		if rr.Code != http.StatusOK {
-			t.Fatalf("unexpected status: got %d want %d body=%q", rr.Code, http.StatusOK, rr.Body.String())
-		}
-	})
-
-	t.Run("wrong password is rejected", func(t *testing.T) {
-		rr := httptest.NewRecorder()
-		req := httptest.NewRequest(http.MethodPost, "/api/v1/login", strings.NewReader(`{"username":"admin","password":"wrong"}`))
-		req.Header.Set("Content-Type", "application/json")
-
-		server.ServeHTTP(rr, req)
-
-		if rr.Code != http.StatusUnauthorized {
-			t.Fatalf("unexpected status: got %d want %d", rr.Code, http.StatusUnauthorized)
-		}
-	})
-}
-
-func TestIngestOnlyKeyScope(t *testing.T) {
-	t.Parallel()
-
-	store := mustOpenStore(t)
-	defer store.Close()
-
-	// Generate a plaintext key and store it as ingest-scope.
-	var raw [32]byte
-	if _, err := rand.Read(raw[:]); err != nil {
-		t.Fatal(err)
-	}
-	plaintext := hex.EncodeToString(raw[:])
-	sum := sha256.Sum256([]byte(plaintext))
-	keySHA256 := hex.EncodeToString(sum[:])
-
-	proj, err := store.ProjectBySlug(context.Background(), "default")
-	if err != nil {
-		t.Fatal(err)
-	}
-	if _, err := store.CreateAPIKey(context.Background(), "test-ingest-key", proj.ID, keySHA256, storage.APIKeyScopeIngest); err != nil {
-		t.Fatal(err)
-	}
-
-	authorizer := auth.New("").WithDBLookup(store.ValidAPIKeySHA256, store.TouchAPIKey)
-	ingestHandler := ingest.NewHandler(authorizer, nil, 1<<20)
-	userAuth, _ := auth.NewUserAuthenticator("admin", "pass", "")
-	sessions := auth.NewSessionManager("secret", time.Hour)
-	server := NewServerWithAuth(ingestHandler, store, userAuth, sessions, nil, nil)
-
-	t.Run("ingest-only key is blocked from protected endpoints", func(t *testing.T) {
-		for _, path := range []string{"/api/v1/issues", "/api/v1/releases", "/api/v1/settings", "/api/v1/apikeys"} {
-			rr := httptest.NewRecorder()
-			req := httptest.NewRequest(http.MethodGet, path, nil)
-			req.Header.Set("x-bugbarn-api-key", plaintext)
-			server.ServeHTTP(rr, req)
-			if rr.Code != http.StatusForbidden {
-				t.Errorf("path %s: got %d want %d", path, rr.Code, http.StatusForbidden)
-			}
-		}
-	})
-
-	t.Run("full-scope key can access protected endpoints", func(t *testing.T) {
-		var rawFull [32]byte
-		if _, err := rand.Read(rawFull[:]); err != nil {
-			t.Fatal(err)
-		}
-		fullPlain := hex.EncodeToString(rawFull[:])
-		sumFull := sha256.Sum256([]byte(fullPlain))
-		if _, err := store.CreateAPIKey(context.Background(), "full-key", proj.ID, hex.EncodeToString(sumFull[:]), storage.APIKeyScopeFull); err != nil {
-			t.Fatal(err)
-		}
-		rr := httptest.NewRecorder()
-		req := httptest.NewRequest(http.MethodGet, "/api/v1/issues", nil)
-		req.Header.Set("x-bugbarn-api-key", fullPlain)
-		server.ServeHTTP(rr, req)
-		if rr.Code != http.StatusOK {
-			t.Errorf("full key /api/v1/issues: got %d want %d body=%q", rr.Code, http.StatusOK, rr.Body.String())
-		}
-	})
-}
-
-func TestReadOnlyKeyScope(t *testing.T) {
-	t.Parallel()
-
-	store := mustOpenStore(t)
-	defer store.Close()
-
-	// Generate a plaintext key and store it as read-scope.
-	var raw [32]byte
-	if _, err := rand.Read(raw[:]); err != nil {
-		t.Fatal(err)
-	}
-	plaintext := hex.EncodeToString(raw[:])
-	sum := sha256.Sum256([]byte(plaintext))
-	keySHA256 := hex.EncodeToString(sum[:])
-
-	proj, err := store.ProjectBySlug(context.Background(), "default")
-	if err != nil {
-		t.Fatal(err)
-	}
-	if _, err := store.CreateAPIKey(context.Background(), "test-read-key", proj.ID, keySHA256, storage.APIKeyScopeRead); err != nil {
-		t.Fatal(err)
-	}
-
-	authorizer := auth.New("").WithDBLookup(store.ValidAPIKeySHA256, store.TouchAPIKey)
-	ingestHandler := ingest.NewHandler(authorizer, nil, 1<<20)
-	userAuth, _ := auth.NewUserAuthenticator("admin", "pass", "")
-	sessions := auth.NewSessionManager("secret", time.Hour)
-	server := NewServerWithAuth(ingestHandler, store, userAuth, sessions, nil, nil)
-
-	t.Run("read-only key can GET protected endpoints", func(t *testing.T) {
-		for _, path := range []string{"/api/v1/issues", "/api/v1/releases", "/api/v1/projects", "/api/v1/logs"} {
-			rr := httptest.NewRecorder()
-			req := httptest.NewRequest(http.MethodGet, path, nil)
-			req.Header.Set("x-bugbarn-api-key", plaintext)
-			server.ServeHTTP(rr, req)
-			if rr.Code != http.StatusOK {
-				t.Errorf("GET %s: got %d want %d body=%q", path, rr.Code, http.StatusOK, rr.Body.String())
-			}
-		}
-	})
-
-	t.Run("read-only key is blocked from POST/PUT/DELETE", func(t *testing.T) {
-		methods := []struct {
-			method string
-			path   string
-		}{
-			{http.MethodPost, "/api/v1/releases"},
-			{http.MethodDelete, "/api/v1/projects/default"},
-			{http.MethodPut, "/api/v1/settings"},
-		}
-		for _, tc := range methods {
-			rr := httptest.NewRecorder()
-			req := httptest.NewRequest(tc.method, tc.path, strings.NewReader("{}"))
-			req.Header.Set("x-bugbarn-api-key", plaintext)
-			req.Header.Set("Content-Type", "application/json")
-			server.ServeHTTP(rr, req)
-			if rr.Code != http.StatusForbidden {
-				t.Errorf("%s %s: got %d want %d", tc.method, tc.path, rr.Code, http.StatusForbidden)
-			}
-		}
-	})
-
-	t.Run("read-only key is blocked from settings and apikeys", func(t *testing.T) {
-		for _, path := range []string{"/api/v1/settings", "/api/v1/apikeys"} {
-			rr := httptest.NewRecorder()
-			req := httptest.NewRequest(http.MethodGet, path, nil)
-			req.Header.Set("x-bugbarn-api-key", plaintext)
-			server.ServeHTTP(rr, req)
-			if rr.Code != http.StatusForbidden {
-				t.Errorf("GET %s: got %d want %d", path, rr.Code, http.StatusForbidden)
-			}
-		}
-	})
-}
-
-func TestIngestCORSHeaders(t *testing.T) {
-	t.Parallel()
-
-	store := mustOpenStore(t)
-	defer store.Close()
-	server := NewServer(nil, store, nil)
-
-	t.Run("OPTIONS preflight returns wildcard ACAO", func(t *testing.T) {
-		rr := httptest.NewRecorder()
-		req := httptest.NewRequest(http.MethodOptions, "/api/v1/events", nil)
-		req.Header.Set("Origin", "https://app.example.com")
-		server.ServeHTTP(rr, req)
-		if rr.Code != http.StatusNoContent {
-			t.Fatalf("got %d want 204", rr.Code)
-		}
-		if got := rr.Header().Get("Access-Control-Allow-Origin"); got != "*" {
-			t.Errorf("ACAO: got %q want %q", got, "*")
-		}
-	})
-
-	t.Run("POST ingest returns wildcard ACAO", func(t *testing.T) {
-		rr := httptest.NewRecorder()
-		req := httptest.NewRequest(http.MethodPost, "/api/v1/events", strings.NewReader(`{}`))
-		req.Header.Set("Origin", "https://app.example.com")
-		req.Header.Set("Content-Type", "application/json")
-		server.ServeHTTP(rr, req)
-		if got := rr.Header().Get("Access-Control-Allow-Origin"); got != "*" {
-			t.Errorf("ACAO on POST: got %q want %q", got, "*")
-		}
-	})
-
-	t.Run("non-ingest endpoint does not get wildcard ACAO", func(t *testing.T) {
-		rr := httptest.NewRecorder()
-		req := httptest.NewRequest(http.MethodGet, "/api/v1/issues", nil)
-		req.Header.Set("Origin", "https://app.example.com")
-		server.ServeHTTP(rr, req)
-		if got := rr.Header().Get("Access-Control-Allow-Origin"); got == "*" {
-			t.Error("expected non-wildcard ACAO on protected endpoint, got *")
 		}
 	})
 }
